third_party/WebKit/LayoutTests/http/tests/security/subresourceIntegrity/subresource-integrity-style-cors-bad-integrity.html

<!DOCTYPE html>
<head>
    <title>SRI on cross origin CORS enabled style, with bad integrity</title>
    <style>.id1 { background-color: red }</style>
    <link rel="stylesheet"
          crossorigin
          integrity="ni:///sha256;deadbeef"
          href="http://localhost:8000/security/resources/cors-style.php"></link>
</head>
<body>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script>
        window.onload = function () {
            var ele = document.getElementById('id1');
            var divBackground = window.getComputedStyle(ele, null).getPropertyValue('background-color');
            assert_equals(divBackground, "rgb(255, 0, 0)");
            done();
        };
    </script>
    <div id="id1" class="id1"></div>
</body>
</html>
