third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/security/subresourceIntegrity/subresource-integrity-style-cors-bad-integrity-expected.txt

CONSOLE ERROR: Failed to find a valid digest in the 'integrity' attribute for resource 'http://localhost:8000/security/resources/cors-style.php' with computed SHA-256 integrity 'fy1lC6jbv7rOG6Flb00n5PYLggqe25IcE6QlZf8KtwI='. The resource has been blocked.
This is a testharness.js-based test.
PASS SRI on cross origin CORS enabled style, with bad integrity 
Harness: the test ran to completion.
